use uint64_t instead of uint32_t in for loops with condition <=
Using uint32_t as index with <= condition may result in overflows when
right side of condition is UINT32_MAX. In such case when the index has
value of UINT32_MAX the condition is stil true, the loop will execute,
and the result of UINT32_MAX + 1 is 0 (overflow) and the condition will
be still true - that results in infinite loop.

+++ src/ledger/LedgerPerformanceTests.cpp
@@ -176,7 +176,8 @@ TEST_CASE("ledger performance test", "[performance][hide]")
     Timer& ledgerTimer = sim.mApp->getMetrics().NewTimer(
         {"performance-test", "ledger", "close"});
     Timer& mergeTimer = sim.mApp->getBucketManager().getMergeTimer();
-    for (int iAccounts = 1000000; iAccounts <= nAccounts; iAccounts *= 10)
+    // use uint64_t for iAccounts to prevent overflows
+    for (uint64_t iAccounts = 1000000; iAccounts <= nAccounts; iAccounts *= 10)
     {
         ledgerTimer.Clear();
         mergeTimer.Clear();
